Update source to beta4.2

+++ shell2/widgets/listview.h
@@ -30,7 +30,7 @@ typedef struct
 	Font *font;
 
 	uint16 colwidth[8];
-	uchar colalign[8];
+	uint16 colalign[8];
 
 	BitMap **icons;
